requirements.txt
asgiref==3.11.0
attrs==25.4.0
beautifulsoup4==4.14.3
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
colorama==0.4.6
coverage==7.13.1
cryptography==46.0.3
dj-database-url==3.1.0
Django==6.0.1
django-filter==25.2
djangorestframework==3.16.1
dotenv==0.9.9
drf-spectacular==0.29.0
gunicorn==23.0.0
idna==3.11
inflection==0.5.1
iniconfig==2.3.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
packaging==25.0
pdfminer.six==20251230
pdfplumber==0.11.9
pillow==12.1.0
pluggy==1.6.0
psycopg2-binary==2.9.11
pycparser==2.23
Pygments==2.19.2
PyPDF2==3.0.1
pypdfium2==5.3.0
pytest==9.0.2
pytest-cov==7.0.0
pytest-django==4.11.1
python-decouple==3.8
python-dotenv==1.2.1
python-slugify==8.0.4
PyYAML==6.0.3
referencing==0.37.0
requests==2.32.5
rpds-py==0.30.0
soupsieve==2.8.1
sqlparse==0.5.5
text-unidecode==1.3
typing_extensions==4.15.0
tzdata==2025.3
uritemplate==4.2.0
urllib3==2.6.2
whitenoise==6.11.0
